UPDATE 滑动修正（？） UPDATE 报告页面初步 需要修改

+++ mr-listen/client/src/components/HoleBar/HoleBar.tsx
@@ -4,12 +4,14 @@ import {ScrollView, View, Text} from '@tarojs/components'
 import Avatar from './../ChatBubble/Avatar'
 
 import './HoleBar.less'
+import Listen from '../../utils/listen';
 
 interface IProp {
   holeAvatarSrc: string,
   holeTitle: string,
   onDelete: () => void,
-  onClick: () => void
+  onClick: () => void,
+  onUpdate: () => void
 }
 
 interface IState {
@@ -31,6 +33,10 @@ export default class HoleBar extends Component<IProp, IState> {
 
   private scrollLeft = 0;
 
+  private DELETE_OPTION = '删除';
+  private UPDATE_OPTION = '修改';
+  private ITEM_LIST = [this.DELETE_OPTION, this.UPDATE_OPTION];
+
   constructor(props) {
     super(props);
     this.state = {
@@ -38,11 +44,38 @@ export default class HoleBar extends Component<IProp, IState> {
     };
   }
 
+  private handleLongPress = () => {
+    Taro.showActionSheet({
+      itemList: this.ITEM_LIST
+    })
+      .then((e) => {
+        if (e.tapIndex >= 0) {
+          if (this.ITEM_LIST[e.tapIndex] === this.DELETE_OPTION) {
+            this.props.onDelete();
+          } else if (this.ITEM_LIST[e.tapIndex] === this.UPDATE_OPTION) {
+            this.props.onUpdate();
+          } else {
+            throw Error(`tapIndex 不匹配 ${e.tapIndex}`);
+          }
+        }
+      })
+      .catch((e) => {
+        const {errMsg = ''} = e;
+        console.error(e);
+        // 是不是取消了
+        if (!errMsg.includes('cancel')) {
+          Listen.message.error('出错了');
+        }
+      })
+  };
+
   private handlerTouchEnd = () => {
     if (this.scrollLeft > this.halfHeight) {
       let scrollLeft = this.height >>> 0;
       this.setState({scrollLeft: scrollLeft - 1}, () => {
-        this.setState({scrollLeft: scrollLeft, isSetting: true});
+        this.setState({scrollLeft: scrollLeft, isSetting: true}, () => {
+          this.forceUpdate();
+        });
       });
     } else {
       this.reset();
@@ -65,7 +98,17 @@ export default class HoleBar extends Component<IProp, IState> {
   private handleDeleteClicked = (e) => {
     if (this.state.isSetting) {
       e.stopPropagation();
-      this.props.onDelete();
+      Taro.showModal({
+        title: '提示',
+        content: `确认删除树洞 ${this.props.holeTitle}`
+      })
+        .then((res) => {
+          if (res.confirm) {
+            this.props.onDelete();
+            this.reset();
+          }
+        })
+        .catch((reason) => console.error(reason));
     }
   };
 
@@ -77,11 +120,11 @@ export default class HoleBar extends Component<IProp, IState> {
 
   render() {
     return (
-      <ScrollView style={{
-        height: this.height + 'px',
-        whiteSpace: 'nowrap',
-        display: 'inline-block'
-      }}
+      <ScrollView className={'hole-bar-scroll-view'}
+                  style={{
+                    height: this.height + 'px'
+                  }}
+                  onLongPress={this.handleLongPress}
                   onTouchEnd={this.handlerTouchEnd}
                   onScroll={this.handleScroll}
                   onClick={this.handleClick}
